cool changes
> Better README
> SIMPLE BENCHMARK: print_benchmark_time can take cpu_frequency as input
> WIN64: win64_alloc can alloc a u64 (removed limit of 2Gb max)

+++ win64_basic.h
@@ -16,7 +16,7 @@ void win64_print_error() {
     print((const char*)buf);
 }
 
-void* win64_alloc(int size) {
+void* win64_alloc(u64 size) {
     return VirtualAlloc(NULL, size, MEM_RESERVE | MEM_COMMIT, PAGE_READWRITE);
 }
